Extract chat marker prompts
